Prevent buffering on server to improve SSE performance
Use single button for playing melds

Import missing library

Remove files from separate projects

Improve script for installing project

+++ setup/server_installation.sh
@@ -9,8 +9,43 @@ fi
 # Prompt for the domain name
 read -p "Enter the domain name (e.g., subdomain.domain.example): " DOMAIN_NAME
 
-# Prompt for the email address for Certbot (optional)
-read -p "Enter your email address for Certbot (this is used for warnings about expiry, deprecation, etc. Leave blank to not provide an email address): " CERTBOT_EMAIL
+# Check if SSL certificate already exists
+if [ ! -f "/etc/letsencrypt/live/$DOMAIN_NAME/fullchain.pem" ]; then
+    is_ssl_certificate_required=true
+else
+    is_ssl_certificate_required=false
+    echo "SSL certificate already exists for $DOMAIN_NAME. Skipping certificate request."
+fi
+
+if [ "$is_ssl_certificate_required" = true ]; then
+    # Prompt for the email address for Certbot (optional)
+    read -p "Enter your email address for Certbot (this is used for warnings about expiry, deprecation, etc. Leave blank to not provide an email address): " CERTBOT_EMAIL
+fi
+
+# Generate SSH Key if it doesn't exist
+if [ ! -f ~/.ssh/id_rsa ]; then
+    ssh-keygen -t rsa -b 4096 -f ~/.ssh/id_rsa -N ""
+fi
+
+# Prompt to add SSH key to GitHub
+cat ~/.ssh/id_rsa.pub
+echo "Add the above SSH key to your GitHub account."
+echo "Press [Enter] key after adding the SSH key to GitHub..."
+read -p ""
+
+# List of GitHub SSH key fingerprints
+GITHUB_KEYS=(
+    "github.com ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIOMqqnkVzrm0SdG6UOoqKLsabgH5C9okWi0dh2l9GKJl"
+    "github.com ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBEmKSENjQEezOmxkZMy7opKgwFB9nkt5YRrYMjNuG5N87uRgg6CLrbo5wAdT/y6v0mKV0U2w0WZ2YB/++Tpockg="
+    "github.com ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQCj7ndNxQowgcQnjshcLrqPEiiphnt+VTTvDP6mHBL9j1aNUkY4Ue1gvwnGLVlOhGeYrnZaMgRK6+PKCUXaDbC7qtbW8gIkhL7aGCsOr/C56SJMy/BCZfxd1nWzAOxSDPgVsmerOBYfNqltV9/hWCqBywINIR+5dIg6JTJ72pcEpEjcYgXkE2YEFXV1JHnsKgbLWNlhScqb2UmyRkQyytRLtL+38TGxkxCflmO+5Z8CSSNY7GidjMIZ7Q4zMjA2n1nGrlTDkzwDCsw+wqFPGQA179cnfGWOWRVruj16z6XyvxvjJwbz0wQZ75XK5tKSb7FNyeIEs4TT4jk+S4dhPeAUC5y+bDYirYgM4GC7uEnztnZyaVWQ7B381AK4Qdrwt51ZqExKbQpTUNn+EjqoTwvqNj4kqx5QUCI0ThS/YkOxJCXmPUWZbhjpCg56i+2aB6CmK2JGhn57K5mj0MNdBXA4/WnwH6XoPWJzK5Nyu2zB3nAZp+S5hpQs+p1vN1/wsjk="
+)
+
+# Add GitHub's SSH key fingerprints to known hosts if not already present
+for KEY in "${GITHUB_KEYS[@]}"; do
+    if ! grep -q "$KEY" ~/.ssh/known_hosts; then
+        echo "$KEY" >> ~/.ssh/known_hosts
+    fi
+done
 
 # Generate a secure MySQL root password excluding specific characters
 MYSQL_ROOT_PASSWORD=$(python3 -c "import secrets, string; print(''.join(secrets.choice(string.ascii_letters + string.digits + '!@#%^&*()_+-=') for _ in range(31)))")
@@ -61,36 +96,20 @@ ALTER USER 'root'@'localhost' IDENTIFIED WITH caching_sha2_password BY '$MYSQL_R
 FLUSH PRIVILEGES;
 MYSQL_SCRIPT
 
-# Generate SSH Key if it doesn't exist
-if [ ! -f ~/.ssh/id_rsa ]; then
-    ssh-keygen -t rsa -b 4096 -f ~/.ssh/id_rsa -N ""
+# Obtain an SSL certificate if required
+if [ "$is_ssl_certificate_required" = true ]; then
+    if [ -z "$CERTBOT_EMAIL" ]; then
+        certbot --nginx -d $DOMAIN_NAME --non-interactive --agree-tos --register-unsafely-without-email
+    else
+        certbot --nginx -d $DOMAIN_NAME --non-interactive --agree-tos -m $CERTBOT_EMAIL
+    fi
 fi
 
-# Prompt to add SSH key to GitHub
-cat ~/.ssh/id_rsa.pub
-echo "Add the above SSH key to your GitHub account."
-echo "Press [Enter] key after adding the SSH key to GitHub..."
-read -p ""
-
-# List of GitHub SSH key fingerprints
-GITHUB_KEYS=(
-    "github.com ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIOMqqnkVzrm0SdG6UOoqKLsabgH5C9okWi0dh2l9GKJl"
-    "github.com ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBEmKSENjQEezOmxkZMy7opKgwFB9nkt5YRrYMjNuG5N87uRgg6CLrbo5wAdT/y6v0mKV0U2w0WZ2YB/++Tpockg="
-    "github.com ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQCj7ndNxQowgcQnjshcLrqPEiiphnt+VTTvDP6mHBL9j1aNUkY4Ue1gvwnGLVlOhGeYrnZaMgRK6+PKCUXaDbC7qtbW8gIkhL7aGCsOr/C56SJMy/BCZfxd1nWzAOxSDPgVsmerOBYfNqltV9/hWCqBywINIR+5dIg6JTJ72pcEpEjcYgXkE2YEFXV1JHnsKgbLWNlhScqb2UmyRkQyytRLtL+38TGxkxCflmO+5Z8CSSNY7GidjMIZ7Q4zMjA2n1nGrlTDkzwDCsw+wqFPGQA179cnfGWOWRVruj16z6XyvxvjJwbz0wQZ75XK5tKSb7FNyeIEs4TT4jk+S4dhPeAUC5y+bDYirYgM4GC7uEnztnZyaVWQ7B381AK4Qdrwt51ZqExKbQpTUNn+EjqoTwvqNj4kqx5QUCI0ThS/YkOxJCXmPUWZbhjpCg56i+2aB6CmK2JGhn57K5mj0MNdBXA4/WnwH6XoPWJzK5Nyu2zB3nAZp+S5hpQs+p1vN1/wsjk="
-)
-
-# Add GitHub's SSH key fingerprints to known hosts if not already present
-for KEY in "${GITHUB_KEYS[@]}"; do
-    if ! grep -q "$KEY" ~/.ssh/known_hosts; then
-        echo "$KEY" >> ~/.ssh/known_hosts
-    fi
-done
-
 # Clone the Repository
-git clone git@github.com:SimonMayer/gin_rummy_2024.git /root/ginrummy
+git clone git@github.com:SimonMayer/gin_rummy_2024.git /root/ginrummy-engine
 
 # Navigate to the App Directory
-cd /root/ginrummy
+cd /root/ginrummy-engine
 
 # Create a Virtual Environment
 python3 -m venv venv
@@ -138,31 +157,24 @@ done
 # Set Up systemd Service
 echo "
 [Unit]
-Description=Gunicorn instance to serve ginrummy
+Description=Gunicorn instance to serve ginrummy-engine
 After=network.target
 
 [Service]
 User=root
 Group=www-data
-WorkingDirectory=/root/ginrummy
-Environment=\"PATH=/root/ginrummy/venv/bin\"
-ExecStart=/root/ginrummy/venv/bin/gunicorn --timeout 3600 --workers 3 --worker-class gevent --bind 127.0.0.1:8000 wsgi:app
+WorkingDirectory=/root/ginrummy-engine
+Environment=\"PATH=/root/ginrummy-engine/venv/bin\"
+ExecStart=/root/ginrummy-engine/venv/bin/gunicorn --timeout 3600 --workers 3 --worker-class gevent --bind 127.0.0.1:8000 wsgi:app
 
 [Install]
 WantedBy=multi-user.target
-" > /etc/systemd/system/ginrummy.service
+" > /etc/systemd/system/ginrummy-engine.service
 
 # Reload systemd and Start Gunicorn Service
 systemctl daemon-reload
-systemctl start ginrummy
-systemctl enable ginrummy
-
-# Obtain an SSL certificate
-if [ -z "$CERTBOT_EMAIL" ]; then
-    certbot --nginx -d $DOMAIN_NAME --non-interactive --agree-tos --register-unsafely-without-email
-else
-    certbot --nginx -d $DOMAIN_NAME --non-interactive --agree-tos -m $CERTBOT_EMAIL
-fi
+systemctl start ginrummy-engine
+systemctl enable ginrummy-engine
 
 # Set Up Nginx as a Reverse Proxy with SSL
 echo "
@@ -191,15 +203,20 @@ server {
         proxy_set_header X-Real-IP \$remote_addr;
         proxy_set_header X-Forwarded-For \$proxy_add_x_forwarded_for;
         proxy_set_header X-Forwarded-Proto \$scheme;
+        proxy_buffering off;
     }
 }
-" > /etc/nginx/sites-available/ginrummy
+" > /etc/nginx/sites-available/ginrummy-engine
 
 # Enable the custom Nginx configuration
-ln -s /etc/nginx/sites-available/ginrummy /etc/nginx/sites-enabled/
+if [ ! -f /etc/nginx/sites-enabled/ginrummy-engine ]; then
+    ln -s /etc/nginx/sites-available/ginrummy-engine /etc/nginx/sites-enabled/
+fi
 
-# Remove the default Nginx configuration
-rm /etc/nginx/sites-enabled/default
+# Remove the default Nginx configuration if it exists
+if [ -f /etc/nginx/sites-enabled/default ]; then
+    rm /etc/nginx/sites-enabled/default
+fi
 
 # Restart Nginx
 systemctl restart nginx
